Don't follow ChatGPT

+++ Sources/BBSwift/SwiftUI/Extensions/Color.swift
@@ -43,15 +43,15 @@ public extension Color {
     init(hex: String) {
         let scanner = Scanner(string: hex)
         scanner.scanLocation = 0
-
+        
         var rgbValue: UInt64 = 0
         scanner.scanHexInt64(&rgbValue)
-
-        let r = CGFloat((rgbValue & 0xFF0000) >> 16) / 255.0
-        let g = CGFloat((rgbValue & 0x00FF00) >> 8) / 255.0
-        let b = CGFloat(rgbValue & 0x0000FF) / 255.0
-
-        self.init(red: Double(r), green: Double(g), blue: Double(b))
+        
+        let r = Double((rgbValue & 0xFF0000) >> 16) / 255.0
+        let g = Double((rgbValue & 0x00FF00) >> 8) / 255.0
+        let b = Double(rgbValue & 0x0000FF) / 255.0
+        
+        self.init(red: r, green: g, blue: b)
     }
 }
 #endif
